backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

psycopg2-binary>=2.9

pandas>=2.0
numpy>=1.24
scikit-learn>=1.3

python-dateutil>=2.8
pytz>=2023.3
